requirements.txt
torch==2.2.2
torchvision==0.17.2
facenet-pytorch
insightface
onnxruntime-gpu
numpy
pillow
tqdm
matplotlib
scikit-learn
faiss-cpu
kagglehub
streamlit
seaborn